feat: implement name resolver
