+++ libs/@local/hashql/ast/src/error.rs
@@ -3,10 +3,14 @@ use alloc::borrow::Cow;
 use hashql_core::span::SpanId;
 use hashql_diagnostics::{Diagnostic, category::DiagnosticCategory};
 
+use crate::lowering::error::LoweringDiagnosticCategory;
+
 pub type AstDiagnostic = Diagnostic<AstDiagnosticCategory, SpanId>;
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub enum AstDiagnosticCategory {}
+pub enum AstDiagnosticCategory {
+    Lowering(LoweringDiagnosticCategory),
+}
 
 impl DiagnosticCategory for AstDiagnosticCategory {
     fn id(&self) -> Cow<'_, str> {
@@ -18,6 +22,8 @@ impl DiagnosticCategory for AstDiagnosticCategory {
     }
 
     fn subcategory(&self) -> Option<&dyn DiagnosticCategory> {
-        match *self {}
+        match self {
+            Self::Lowering(lowering) => Some(lowering),
+        }
     }
 }
