+++ libs/@local/hashql/ast/src/lowering/error.rs
@@ -3,8 +3,12 @@ use alloc::borrow::Cow;
 use hashql_core::span::SpanId;
 use hashql_diagnostics::{Diagnostic, category::DiagnosticCategory};
 
-pub(crate) type LoweringDiagnostic = Diagnostic<LoweringDiagnosticCategory, SpanId>;
+pub type LoweringDiagnostic = Diagnostic<LoweringDiagnosticCategory, SpanId>;
 
+#[expect(
+    clippy::empty_enum,
+    reason = "Preparation for future diagnostic categories"
+)]
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
 pub enum LoweringDiagnosticCategory {}
 
@@ -18,6 +22,6 @@ impl DiagnosticCategory for LoweringDiagnosticCategory {
     }
 
     fn subcategory(&self) -> Option<&dyn DiagnosticCategory> {
-        match *self {}
+        None
     }
 }
