+++ libs/@local/hashql/ast/src/lowering/mod.rs
@@ -1,2 +1,2 @@
 pub mod error;
-pub mod resolve_names;
+pub mod name_resolver;
